Fixing concurrency issue between consumer and authenticator components
The authenticator component was expecting the consumer to move any DB in "creating" state to "ready" state, which happens when the consume consumes a message for that DB/topic.

In a multinode setup, it is possible that the consumer would never see a message for a particular topic, hence the authenticator will not realize that the db is ready.

Fixing the issue by treating the auth and consumer components as completely separate entities that do not share any in-memory state, as it is expected will happen whne we decide to split them into two different processes.

For now, this means that the authenticator always goes to the master db to know the db credentials of a tenant. If we decide to cache this, the cache should be completely independent of any consumer state.